--- app/(auth)/(routes)/sign-up/[[...sign-up]]/page.tsx
@@ -1,5 +1,0 @@
-import { SignUp } from "@clerk/nextjs";
-
-export default function Page() {
-  return <SignUp />;
-}
