+++ components/Providers.tsx
@@ -1,28 +1,28 @@
 "use client";
 
-import { trpc } from "@/app/_trpc/client";
+import { trpc } from "@/src/app/[locale]/_trpc/client";
 import { absoluteUrl } from "@/lib/utils";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { httpBatchLink } from "@trpc/client";
 import { PropsWithChildren, useState } from "react";
 
 const Providers = ({ children }: PropsWithChildren) => {
-  const [queryClient] = useState(() => new QueryClient());
-  const [trpcClient] = useState(() =>
-    trpc.createClient({
-      links: [
-        httpBatchLink({
-          url: absoluteUrl("/api/trpc"),
-        }),
-      ],
-    })
-  );
+	const [queryClient] = useState(() => new QueryClient());
+	const [trpcClient] = useState(() =>
+		trpc.createClient({
+			links: [
+				httpBatchLink({
+					url: absoluteUrl("/api/trpc"),
+				}),
+			],
+		})
+	);
 
-  return (
-    <trpc.Provider client={trpcClient} queryClient={queryClient}>
-      <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
-    </trpc.Provider>
-  );
+	return (
+		<trpc.Provider client={trpcClient} queryClient={queryClient}>
+			<QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
+		</trpc.Provider>
+	);
 };
 
 export default Providers;
