+++ components/chat/chat-context.tsx
@@ -1,209 +1,208 @@
 import { ReactNode, createContext, useRef, useState } from "react";
 import { toast } from "react-hot-toast";
 import { useMutation } from "@tanstack/react-query";
-import { trpc } from "@/app/_trpc/client";
+import { trpc } from "@/src/app/[locale]/_trpc/client";
 import { MAX_FREE_COUNTS } from "@/constant";
 import { v4 as uuidv4 } from "uuid";
 
 type StreamResponse = {
-  addMessage: () => void;
-  message: string;
-  handleInputChange: (event: React.ChangeEvent<HTMLTextAreaElement>) => void;
-  isLoading: boolean;
+	addMessage: () => void;
+	message: string;
+	handleInputChange: (event: React.ChangeEvent<HTMLTextAreaElement>) => void;
+	isLoading: boolean;
 };
 
 export const ChatContext = createContext<StreamResponse>({
-  addMessage: () => {},
-  message: "",
-  handleInputChange: () => {},
-  isLoading: false,
+	addMessage: () => {},
+	message: "",
+	handleInputChange: () => {},
+	isLoading: false,
 });
 
 interface Props {
-  fileId: string;
-  children: ReactNode;
+	fileId: string;
+	children: ReactNode;
 }
 
 export const ChatContextProvider = ({ fileId, children }: Props) => {
-  const [message, setMessage] = useState<string>("");
-  const [isLoading, setIsLoading] = useState<boolean>(false);
+	const [message, setMessage] = useState<string>("");
+	const [isLoading, setIsLoading] = useState<boolean>(false);
 
-  const utils = trpc.useUtils();
+	const utils = trpc.useUtils();
 
-  const backupMessage = useRef("");
+	const backupMessage = useRef("");
 
-  const { mutate: sendMessage } = useMutation({
-    mutationFn: async ({ message }: { message: string }) => {
-      const response = await fetch("/api/message", {
-        method: "POST",
-        body: JSON.stringify({
-          fileId,
-          message,
-        }),
-      });
+	const { mutate: sendMessage } = useMutation({
+		mutationFn: async ({ message }: { message: string }) => {
+			const response = await fetch("/api/message", {
+				method: "POST",
+				body: JSON.stringify({
+					fileId,
+					message,
+				}),
+			});
 
-      if (!response.ok) {
-        throw new Error("Failed to send message");
-      }
+			if (!response.ok) {
+				throw new Error("Failed to send message");
+			}
 
-      return response.body;
-    },
-    networkMode: "always",
-    retry: 3,
+			return response.body;
+		},
+		networkMode: "always",
+		retry: 3,
 
-    onMutate: async ({ message }) => {
-      backupMessage.current = message;
-      setMessage("");
+		onMutate: async ({ message }) => {
+			backupMessage.current = message;
+			setMessage("");
 
-      // step 1
-      await utils.getFileMessages.cancel();
+			// step 1
+			await utils.getFileMessages.cancel();
 
-      // step 2
-      const previousMessages = utils.getFileMessages.getInfiniteData();
+			// step 2
+			const previousMessages = utils.getFileMessages.getInfiniteData();
 
-      // step 3
-      utils.getFileMessages.setInfiniteData(
-        { fileId, limit: MAX_FREE_COUNTS },
-        (old) => {
-          if (!old) {
-            return {
-              pages: [],
-              pageParams: [],
-            };
-          }
+			// step 3
+			utils.getFileMessages.setInfiniteData(
+				{ fileId, limit: MAX_FREE_COUNTS },
+				(old) => {
+					if (!old) {
+						return {
+							pages: [],
+							pageParams: [],
+						};
+					}
 
-          let newPages = [...old.pages];
+					let newPages = [...old.pages];
 
-          let latestPage = newPages[0]!;
+					let latestPage = newPages[0]!;
 
-          latestPage.messages = [
-            {
-              createdAt: new Date().toISOString(),
-              id: uuidv4(),
-              text: message,
-              isUserMessage: true,
-            },
-            ...latestPage.messages,
-          ];
+					latestPage.messages = [
+						{
+							createdAt: new Date().toISOString(),
+							id: uuidv4(),
+							text: message,
+							isUserMessage: true,
+						},
+						...latestPage.messages,
+					];
 
-          newPages[0] = latestPage;
+					newPages[0] = latestPage;
 
-          return {
-            ...old,
-            pages: newPages,
-          };
-        }
-      );
+					return {
+						...old,
+						pages: newPages,
+					};
+				}
+			);
 
-      setIsLoading(true);
+			setIsLoading(true);
 
-      return {
-        previousMessages:
-          previousMessages?.pages.flatMap((page) => page.messages) ?? [],
-      };
-    },
-    onSuccess: async (stream) => {
-      setIsLoading(false);
+			return {
+				previousMessages:
+					previousMessages?.pages.flatMap((page) => page.messages) ?? [],
+			};
+		},
+		onSuccess: async (stream) => {
+			setIsLoading(false);
 
-      if (!stream) {
-        return toast("There was a problem sending this message");
-      }
+			if (!stream) {
+				return toast("There was a problem sending this message");
+			}
 
-      const reader = stream.getReader();
-      const decoder = new TextDecoder();
-      let done = false;
+			const reader = stream.getReader();
+			const decoder = new TextDecoder();
+			let done = false;
 
-      // accumulated response
-      let accResponse = "";
+			// accumulated response
+			let accResponse = "";
 
-      while (!done) {
-        const { value, done: doneReading } = await reader.read();
-        done = doneReading;
-        const chunkValue = decoder.decode(value);
+			while (!done) {
+				const { value, done: doneReading } = await reader.read();
+				done = doneReading;
+				const chunkValue = decoder.decode(value);
 
-        accResponse += chunkValue;
+				accResponse += chunkValue;
 
-        // append chunk to the actual message
-        utils.getFileMessages.setInfiniteData(
-          { fileId, limit: MAX_FREE_COUNTS },
-          (old) => {
-            if (!old) return { pages: [], pageParams: [] };
+				// append chunk to the actual message
+				utils.getFileMessages.setInfiniteData(
+					{ fileId, limit: MAX_FREE_COUNTS },
+					(old) => {
+						if (!old) return { pages: [], pageParams: [] };
 
-            let isAiResponseCreated = old.pages.some((page) =>
-              page.messages.some((message) => message.id === "ai-response")
-            );
+						let isAiResponseCreated = old.pages.some((page) =>
+							page.messages.some((message) => message.id === "ai-response")
+						);
 
-            let updatedPages = old.pages.map((page) => {
-              if (page === old.pages[0]) {
-                let updatedMessages;
+						let updatedPages = old.pages.map((page) => {
+							if (page === old.pages[0]) {
+								let updatedMessages;
 
-                if (!isAiResponseCreated) {
-                  updatedMessages = [
-                    {
-                      createdAt: new Date().toISOString(),
-                      id: "ai-response",
-                      text: accResponse,
-                      isUserMessage: false,
-                    },
-                    ...page.messages,
-                  ];
-                } else {
-                  updatedMessages = page.messages.map((message) => {
-                    if (message.id === "ai-response") {
-                      return {
-                        ...message,
-                        text: accResponse,
-                      };
-                    }
-                    return message;
-                  });
-                }
+								if (!isAiResponseCreated) {
+									updatedMessages = [
+										{
+											createdAt: new Date().toISOString(),
+											id: "ai-response",
+											text: accResponse,
+											isUserMessage: false,
+										},
+										...page.messages,
+									];
+								} else {
+									updatedMessages = page.messages.map((message) => {
+										if (message.id === "ai-response") {
+											return {
+												...message,
+												text: accResponse,
+											};
+										}
+										return message;
+									});
+								}
 
-                return {
-                  ...page,
-                  messages: updatedMessages,
-                };
-              }
+								return {
+									...page,
+									messages: updatedMessages,
+								};
+							}
 
-              return page;
-            });
+							return page;
+						});
 
-            return { ...old, pages: updatedPages };
-          }
-        );
-      }
-    },
+						return { ...old, pages: updatedPages };
+					}
+				);
+			}
+		},
 
-    onError: (_, __, context) => {
-      setMessage(backupMessage.current);
-      utils.getFileMessages.setData(
-        { fileId },
-        { messages: context?.previousMessages ?? [] }
-      );
-    },
-    onSettled: async () => {
-      setIsLoading(false);
+		onError: (_, __, context) => {
+			setMessage(backupMessage.current);
+			utils.getFileMessages.setData(
+				{ fileId },
+				{ messages: context?.previousMessages ?? [] }
+			);
+		},
+		onSettled: async () => {
+			setIsLoading(false);
 
-      await utils.getFileMessages.invalidate({ fileId });
-    },
-  });
+			await utils.getFileMessages.invalidate({ fileId });
+		},
+	});
 
-  const handleInputChange = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
-    setMessage(e.target.value);
-  };
+	const handleInputChange = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
+		setMessage(e.target.value);
+	};
 
-  const addMessage = () => sendMessage({ message });
+	const addMessage = () => sendMessage({ message });
 
-  return (
-    <ChatContext.Provider
-      value={{
-        addMessage,
-        message,
-        handleInputChange,
-        isLoading,
-      }}
-    >
-      {children}
-    </ChatContext.Provider>
-  );
+	return (
+		<ChatContext.Provider
+			value={{
+				addMessage,
+				message,
+				handleInputChange,
+				isLoading,
+			}}>
+			{children}
+		</ChatContext.Provider>
+	);
 };
