+++ components/chat/chat-wrapper.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { trpc } from "@/app/_trpc/client";
+import { trpc } from "@/src/app/[locale]/_trpc/client";
 import ChatInput from "./chat-input";
 import Messages from "./messages";
 import { ChevronLeft, Loader2, XCircle } from "lucide-react";
@@ -10,100 +10,99 @@ import { ChatContextProvider } from "./chat-context";
 // import { PLANS } from '@/config/stripe'
 
 interface ChatWrapperProps {
-  fileId: string;
-  isSubscribed: boolean;
+	fileId: string;
+	isSubscribed: boolean;
 }
 
 const ChatWrapper = ({ fileId, isSubscribed }: ChatWrapperProps) => {
-  const { data, isLoading } = trpc.getFileUploadStatus.useQuery(
-    {
-      fileId,
-    },
-    {
-      refetchInterval: (data) =>
-        data?.status === "SUCCESS" || data?.status === "FAILED" ? false : 500,
-      networkMode: "always",
-    }
-  );
+	const { data, isLoading } = trpc.getFileUploadStatus.useQuery(
+		{
+			fileId,
+		},
+		{
+			refetchInterval: (data) =>
+				data?.status === "SUCCESS" || data?.status === "FAILED" ? false : 500,
+			networkMode: "always",
+		}
+	);
 
-  if (isLoading)
-    return (
-      <div className="relative min-h-full bg-zinc-50 flex divide-y divide-zinc-200 flex-col justify-between gap-2">
-        <div className="flex-1 flex justify-center items-center flex-col mb-28">
-          <div className="flex flex-col items-center gap-2">
-            <Loader2 className="h-8 w-8 text-blue-500 animate-spin" />
-            <h3 className="font-semibold text-xl">Loading...</h3>
-            <p className="text-zinc-500 text-sm">
-              We&apos;re preparing your PDF.
-            </p>
-          </div>
-        </div>
+	if (isLoading)
+		return (
+			<div className="relative min-h-full bg-zinc-50 flex divide-y divide-zinc-200 flex-col justify-between gap-2">
+				<div className="flex-1 flex justify-center items-center flex-col mb-28">
+					<div className="flex flex-col items-center gap-2">
+						<Loader2 className="h-8 w-8 text-blue-500 animate-spin" />
+						<h3 className="font-semibold text-xl">Loading...</h3>
+						<p className="text-zinc-500 text-sm">
+							We&apos;re preparing your PDF.
+						</p>
+					</div>
+				</div>
 
-        <ChatInput isDisabled />
-      </div>
-    );
+				<ChatInput isDisabled />
+			</div>
+		);
 
-  if (data?.status === "PROCESSING")
-    return (
-      <div className="relative min-h-full bg-zinc-50 flex divide-y divide-zinc-200 flex-col justify-between gap-2">
-        <div className="flex-1 flex justify-center items-center flex-col mb-28">
-          <div className="flex flex-col items-center gap-2">
-            <Loader2 className="h-8 w-8 text-blue-500 animate-spin" />
-            <h3 className="font-semibold text-xl">Processing PDF...</h3>
-            <p className="text-zinc-500 text-sm">This won&apos;t take long.</p>
-          </div>
-        </div>
+	if (data?.status === "PROCESSING")
+		return (
+			<div className="relative min-h-full bg-zinc-50 flex divide-y divide-zinc-200 flex-col justify-between gap-2">
+				<div className="flex-1 flex justify-center items-center flex-col mb-28">
+					<div className="flex flex-col items-center gap-2">
+						<Loader2 className="h-8 w-8 text-blue-500 animate-spin" />
+						<h3 className="font-semibold text-xl">Processing PDF...</h3>
+						<p className="text-zinc-500 text-sm">This won&apos;t take long.</p>
+					</div>
+				</div>
 
-        <ChatInput isDisabled />
-      </div>
-    );
+				<ChatInput isDisabled />
+			</div>
+		);
 
-  if (data?.status === "FAILED")
-    return (
-      <div className="relative min-h-full bg-zinc-50 flex divide-y divide-zinc-200 flex-col justify-between gap-2">
-        <div className="flex-1 flex justify-center items-center flex-col mb-28">
-          <div className="flex flex-col items-center gap-2">
-            <XCircle className="h-8 w-8 text-red-500" />
-            <h3 className="font-semibold text-xl">Too many pages in PDF</h3>
-            <p className="text-zinc-500 text-sm">
-              Your{" "}
-              <span className="font-medium">
-                {isSubscribed ? "Premium" : "Free"}
-              </span>{" "}
-              plan supports up to{" "}
-              {/* {isSubscribed
+	if (data?.status === "FAILED")
+		return (
+			<div className="relative min-h-full bg-zinc-50 flex divide-y divide-zinc-200 flex-col justify-between gap-2">
+				<div className="flex-1 flex justify-center items-center flex-col mb-28">
+					<div className="flex flex-col items-center gap-2">
+						<XCircle className="h-8 w-8 text-red-500" />
+						<h3 className="font-semibold text-xl">Too many pages in PDF</h3>
+						<p className="text-zinc-500 text-sm">
+							Your{" "}
+							<span className="font-medium">
+								{isSubscribed ? "Premium" : "Free"}
+							</span>{" "}
+							plan supports up to{" "}
+							{/* {isSubscribed
                 ? PLANS.find((p) => p.name === "Premium")?.pagesPerPdf
                 : PLANS.find((p) => p.name === "Free")?.pagesPerPdf}{" "}
               pages per PDF. */}
-            </p>
-            <Link
-              href="/dashboard"
-              className={buttonVariants({
-                variant: "secondary",
-                className: "mt-4",
-              })}
-            >
-              <ChevronLeft className="h-3 w-3 mr-1.5" />
-              Back
-            </Link>
-          </div>
-        </div>
+						</p>
+						<Link
+							href="/dashboard"
+							className={buttonVariants({
+								variant: "secondary",
+								className: "mt-4",
+							})}>
+							<ChevronLeft className="h-3 w-3 mr-1.5" />
+							Back
+						</Link>
+					</div>
+				</div>
 
-        <ChatInput isDisabled />
-      </div>
-    );
+				<ChatInput isDisabled />
+			</div>
+		);
 
-  return (
-    <ChatContextProvider fileId={fileId}>
-      <div className="relative min-h-full bg-zinc-50 flex divide-y divide-zinc-200 flex-col justify-between gap-2">
-        <div className="flex-1 justify-between flex flex-col mb-28">
-          <Messages fileId={fileId} />
-        </div>
+	return (
+		<ChatContextProvider fileId={fileId}>
+			<div className="relative min-h-full bg-zinc-50 flex divide-y divide-zinc-200 flex-col justify-between gap-2">
+				<div className="flex-1 justify-between flex flex-col mb-28">
+					<Messages fileId={fileId} />
+				</div>
 
-        <ChatInput />
-      </div>
-    </ChatContextProvider>
-  );
+				<ChatInput />
+			</div>
+		</ChatContextProvider>
+	);
 };
 
 export default ChatWrapper;
