+++ lib/upload-thing.ts
@@ -1,5 +1,5 @@
 import { generateReactHelpers } from "@uploadthing/react/hooks";
 
-import type { OurFileRouter } from "@/app/api/uploadthing/core";
+import type { OurFileRouter } from "@/src/app/[locale]/api/uploadthing/core";
 
 export const { useUploadThing } = generateReactHelpers<OurFileRouter>();
