+++ messages/en.json
@@ -1,0 +1,5 @@
+{
+	"landing": {
+		"btn-start": "Get Started"
+	}
+}
