+++ package.json
@@ -37,8 +37,8 @@
     "@trpc/react-query": "^10.44.1",
     "@trpc/server": "^10.44.1",
     "@types/node": "20.4.4",
-    "@types/react": "18.2.15",
-    "@types/react-dom": "18.2.7",
+    "@types/react": "^18.2.58",
+    "@types/react-dom": "^18.2.19",
     "@uploadthing/react": "^6.2.3",
     "ai": "^2.2.27",
     "autoprefixer": "10.4.14",
@@ -49,19 +49,20 @@
     "d3-dsv": "^2.0.0",
     "date-fns": "^3.0.0-alpha.2",
     "eslint": "8.45.0",
-    "eslint-config-next": "13.4.12",
+    "eslint-config-next": "^14.1.0",
     "immer": "^10.0.3",
     "langchain": "^0.0.200",
     "lucide-react": "^0.334.0",
     "mammoth": "^1.6.0",
     "midtrans-client": "^1.3.1",
-    "next": "13.4.12",
+    "next": "^14.1.0",
+    "next-intl": "^3.9.1",
     "openai": "^4.11.1",
     "papaparse": "^5.4.1",
     "pdf-parse": "^1.1.1",
     "postcss": "8.4.27",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-dropzone": "^14.2.3",
     "react-hook-form": "^7.45.2",
     "react-hot-toast": "^2.4.1",
