+++ src/app/[locale]/(auth)/(routes)/sign-in/[[...sign-in]]/page.tsx
@@ -1,0 +1,11 @@
+import { SignIn } from "@clerk/nextjs";
+
+export default function Page({
+	params: { locale },
+}: {
+	params: { locale: string };
+}) {
+	return (
+		<SignIn path={`/${locale}/sign-in`} redirectUrl={`/${locale}/dashboard`} />
+	);
+}
