+++ src/app/[locale]/(auth)/(routes)/sign-up/[[...sign-up]]/page.tsx
@@ -1,0 +1,11 @@
+import { SignUp } from "@clerk/nextjs";
+
+export default function Page({
+	params: { locale },
+}: {
+	params: { locale: string };
+}) {
+	return (
+		<SignUp path={`/${locale}/sign-up`} redirectUrl={`/${locale}/dashboard`} />
+	);
+}
