+++ src/app/[locale]/(dashboard)/(routes)/bank-statement-analyzer/page.tsx
