+++ src/app/[locale]/(dashboard)/(routes)/code/constant.ts
