+++ src/app/[locale]/(dashboard)/(routes)/code/page.tsx
