+++ src/app/[locale]/(dashboard)/(routes)/coming-soon/page.tsx
