+++ src/app/[locale]/(dashboard)/(routes)/contract-checker/page.tsx
