+++ src/app/[locale]/(dashboard)/(routes)/cv-scanner/[fileid]/page.tsx
