+++ src/app/[locale]/(dashboard)/(routes)/dashboard/page.tsx
