+++ src/app/[locale]/(dashboard)/(routes)/help-center/page.tsx
