+++ src/app/[locale]/(dashboard)/(routes)/history/page.tsx
