+++ src/app/[locale]/(dashboard)/(routes)/loan-application-processor/page.tsx
