+++ src/app/[locale]/(dashboard)/(routes)/pricing/page.tsx
