+++ src/app/[locale]/(dashboard)/(routes)/regulatory-auditor/page.tsx
