+++ src/app/[locale]/(landing)/error.tsx
