+++ src/app/[locale]/(landing)/layout.tsx
