+++ src/app/[locale]/api/cv-analyzer/route.ts
