+++ src/app/[locale]/api/update-user-tos/route.ts
