+++ src/app/[locale]/api/uploadthing/core.ts
